backend/services/auth_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.25
asyncpg==0.29.0
python-multipart==0.0.6
httpx==0.26.0
email-validator==2.1.0
